move `index.ts` to `parser/main.ts`
