+++ lib/combinators/lazy-evaluation/memoize.ts
@@ -1,4 +1,4 @@
-import { createParser, type Parser } from "../../../src/index.ts";
+import { createParser, type Parser } from "../../../src/parser/main.ts";
 
 /**
  * Takes a parser thunk and memoize it upon evaluation.
