+++ lib/combinators/sequencing/last.ts
@@ -1,5 +1,5 @@
 import { result } from "../../primitives/result.ts";
-import type { Parser } from "../../../src/index.ts";
+import type { Parser } from "../../../src/parser/main.ts";
 
 /**
  * Skips the first parser
