+++ lib/tokens/integer.ts
@@ -1,5 +1,5 @@
 import { parseErrors } from "../../src/errors.ts";
-import type { Parser } from "../../src/index.ts";
+import type { Parser } from "../../src/parser/main.ts";
 import { alt } from "../combinators/choice/alt.ts";
 import { literal } from "./literal.ts";
 import { natural } from "./natural.ts";
