+++ lib/tokens/letter.ts
@@ -1,5 +1,5 @@
 import { parseErrors } from "../../src/errors.ts";
-import type { Parser } from "../../src/index.ts";
+import type { Parser } from "../../src/parser/main.ts";
 import { regex } from "./regex.ts";
 
 /**
