+++ lib/tokens/letters.ts
@@ -1,4 +1,4 @@
-import type { Parser } from "../../src/index.ts";
+import type { Parser } from "../../src/parser/main.ts";
 import { many } from "../combinators/iteration/many.ts";
 import { letter } from "./letter.ts";
